+++ src/features/project/new/ui/application/QuestionCard.tsx
@@ -2,11 +2,11 @@ import { useSortable } from "@dnd-kit/sortable"
 import { CSS } from "@dnd-kit/utilities"
 import { useRef, useState } from "react"
 
-import { CheckboxFieldList } from "@/shared/ui/input/checkbox/CheckboxFieldList"
-import { FileUploadField } from "@/shared/ui/input/file/FileUploadField"
-import { PortfolioField } from "@/shared/ui/input/file/PortfolioField"
-import { RadioFieldList } from "@/shared/ui/input/radio/RadioFieldList"
+import { CheckboxFieldList } from "@/shared/ui/question-field/CheckboxFieldList"
+import { FileUploadField } from "@/shared/ui/question-field/FileUploadField"
+import { PortfolioField } from "@/shared/ui/question-field/PortfolioField"
 import { QuestionForm } from "@/shared/ui/question-field/QuestionForm"
+import { RadioFieldList } from "@/shared/ui/question-field/RadioFieldList"
 import { TextQuestionField } from "@/shared/ui/question-field/TextQuestionField"
 
 import type { Question } from "@/features/project/new/model/applicationQuestion"
